requirements.txt
captcha==0.4
certifi==2023.7.22
charset-normalizer==2.0.12
cycler==0.11.0
dataclasses==0.8
idna==3.4
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.4.0
pyparsing==3.1.0
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
torch @ file:///D:/PycharmProjects/DeepLearningImageCaptcha/lib/torch-1.10.2%2Bcu113-cp36-cp36m-win_amd64.whl
torchvision @ file:///D:/PycharmProjects/DeepLearningImageCaptcha/lib/torchvision-0.11.3%2Bcu113-cp36-cp36m-win_amd64.whl
typing_extensions==4.1.1
urllib3==1.26.16
